refactor to use kotlin flows

+++ app/src/main/java/com/tommunyiri/dvtweatherapp/presentation/screens/search/SearchScreenState.kt
@@ -11,5 +11,5 @@ data class SearchScreenState(
     val weather: Weather? = null,
     val isLoading: Boolean = false,
     val error: String? = null,
-    val addToFavoriteResult: Int? = null,
+    val addToFavoriteResult: Long? = null,
 )
